backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
yfinance==0.2.28
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
supabase==2.3.0
google-generativeai==0.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
aiofiles==23.2.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
email-validator==2.1.0